Refactored hooks and fixes some minor issues
